chore: move tests into own folder
