+++ tests/integration/tests/email/drought/email-uga-drought.test.ts
@@ -1,13 +1,13 @@
 import { JSDOM } from 'jsdom';
 
-import { DisasterType } from '../../../../services/API-service/src/api/disaster/disaster-type.enum';
+import { DisasterType } from '../../../../../services/API-service/src/api/disaster/disaster-type.enum';
 import {
   getAccessToken,
   getEventTitle,
   mockDynamicData,
   resetDB,
   sendNotification,
-} from '../../helpers/utility.helper';
+} from '../../../helpers/utility.helper';
 
 const countryCodeISO3 = 'UGA';
 const disasterType = DisasterType.Drought;
