+++ tests/integration/tests/email/flash-flood/test-flash-flood-scenario.helper.ts
@@ -1,26 +1,29 @@
 import { JSDOM } from 'jsdom';
-
-import { DisasterType } from '../../../../services/API-service/src/api/disaster/disaster-type.enum';
-import { MalariaScenario } from '../../../../services/API-service/src/scripts/enum/mock-scenario.enum';
+import { DisasterType } from '../../../../../services/API-service/src/api/disaster/disaster-type.enum';
+import { FlashFloodsScenario } from '../../../../../services/API-service/src/scripts/enum/mock-scenario.enum';
 import {
   getEventTitle,
-  mockMalaria,
+  mockFlashFlood,
   sendNotification,
-} from '../../helpers/utility.helper';
+} from '../../../helpers/utility.helper';
 
-export async function testMalariaScenario(
-  scenario: MalariaScenario,
+export async function testFlashFloodScenario(
+  scenario: FlashFloodsScenario,
   countryCodeISO3: string,
+  eventNames: string[] = [],
   accessToken: string,
 ): Promise<boolean> {
-  const eventNames = ['0-month', '1-month', '2-month'];
-  const disasterTypeLabel = DisasterType.Malaria;
+  const disasterTypeLabel = 'Flash Flood'; // DisasterType.FlashFloods does not match
 
-  const mockResult = await mockMalaria(scenario, countryCodeISO3, accessToken);
+  const mockResult = await mockFlashFlood(
+    scenario,
+    countryCodeISO3,
+    accessToken,
+  );
   // Act
   const response = await sendNotification(
     countryCodeISO3,
-    DisasterType.Malaria,
+    DisasterType.FlashFloods,
     accessToken,
   );
   // Assert
@@ -28,10 +31,13 @@ export async function testMalariaScenario(
   expect(mockResult.status).toBe(202);
   expect(response.status).toBe(201);
 
-  if (scenario === MalariaScenario.Trigger) {
-    expect(response.body.activeEvents.email).toBeDefined();
-  } else {
+  if (
+    scenario === FlashFloodsScenario.NoTrigger ||
+    scenario.startsWith('trigger-ongoing-') // ongoing triggers are not listed in emails
+  ) {
     expect(response.body.activeEvents.email).toBeUndefined();
+  } else {
+    expect(response.body.activeEvents.email).toBeDefined();
   }
 
   expect(response.body.activeEvents.whatsapp).toBeFalsy();
@@ -47,13 +53,13 @@ export async function testMalariaScenario(
     (el) => (el as Element).textContent.toLowerCase(),
   ).map((el) => el.trim());
 
-  if (scenario === MalariaScenario.Trigger) {
-    expect(eventNamesInEmail.length).toBe(eventNames.length);
-  } else {
+  if (scenario === FlashFloodsScenario.NoTrigger) {
     expect(eventNamesInEmail.length).toBe(0);
+  } else {
+    expect(eventNamesInEmail.length).toBe(eventNames.length);
   }
 
-  if (scenario === MalariaScenario.Trigger) {
+  if (scenario !== FlashFloodsScenario.NoTrigger) {
     // Check if each expected event name is included in at least one title
     for (const eventName of eventNames) {
       const eventTitle = getEventTitle(disasterTypeLabel, eventName);
