+++ tests/integration/tests/email/floods/test-flood-scenario.helper.ts
@@ -1,14 +1,14 @@
 import { JSDOM } from 'jsdom';
 
-import { DisasterType } from '../../../../services/API-service/src/api/disaster/disaster-type.enum';
-import { FloodsScenario } from '../../../../services/API-service/src/scripts/enum/mock-scenario.enum';
+import { DisasterType } from '../../../../../services/API-service/src/api/disaster/disaster-type.enum';
+import { FloodsScenario } from '../../../../../services/API-service/src/scripts/enum/mock-scenario.enum';
 
-import disasters from '../../../../services/API-service/src/scripts/json/disasters.json';
+import disasters from '../../../../../services/API-service/src/scripts/json/disasters.json';
 import {
   getEventTitle,
   mockFloods,
   sendNotification,
-} from '../../helpers/utility.helper';
+} from '../../../helpers/utility.helper';
 
 interface Event {
   eventName: string;
