+++ tests/integration/tests/email/malaria/email-eth-malaria.test.ts
@@ -1,5 +1,5 @@
-import { MalariaScenario } from '../../../../services/API-service/src/scripts/enum/mock-scenario.enum';
-import { getAccessToken, resetDB } from '../../helpers/utility.helper';
+import { MalariaScenario } from '../../../../../services/API-service/src/scripts/enum/mock-scenario.enum';
+import { getAccessToken, resetDB } from '../../../helpers/utility.helper';
 import { testMalariaScenario } from './test-malaria-scenario.helper';
 
 const countryCodeISO3 = 'ETH';
