+++ tests/integration/tests/email/malaria/test-malaria-scenario.helper.ts
@@ -1,29 +1,25 @@
 import { JSDOM } from 'jsdom';
 import { DisasterType } from '../../../../services/API-service/src/api/disaster/disaster-type.enum';
-import { FlashFloodsScenario } from '../../../../services/API-service/src/scripts/enum/mock-scenario.enum';
+import { MalariaScenario } from '../../../../services/API-service/src/scripts/enum/mock-scenario.enum';
 import {
   getEventTitle,
-  mockFlashFlood,
+  mockMalaria,
   sendNotification,
-} from '../../helpers/utility.helper';
+} from '../../../helpers/utility.helper';
 
-export async function testFlashFloodScenario(
-  scenario: FlashFloodsScenario,
+export async function testMalariaScenario(
+  scenario: MalariaScenario,
   countryCodeISO3: string,
-  eventNames: string[] = [],
   accessToken: string,
 ): Promise<boolean> {
-  const disasterTypeLabel = 'Flash Flood'; // DisasterType.FlashFloods does not match
+  const eventNames = ['0-month', '1-month', '2-month'];
+  const disasterTypeLabel = DisasterType.Malaria;
 
-  const mockResult = await mockFlashFlood(
-    scenario,
-    countryCodeISO3,
-    accessToken,
-  );
+  const mockResult = await mockMalaria(scenario, countryCodeISO3, accessToken);
   // Act
   const response = await sendNotification(
     countryCodeISO3,
-    DisasterType.FlashFloods,
+    DisasterType.Malaria,
     accessToken,
   );
   // Assert
@@ -31,13 +27,10 @@ export async function testFlashFloodScenario(
   expect(mockResult.status).toBe(202);
   expect(response.status).toBe(201);
 
-  if (
-    scenario === FlashFloodsScenario.NoTrigger ||
-    scenario.startsWith('trigger-ongoing-') // ongoing triggers are not listed in emails
-  ) {
-    expect(response.body.activeEvents.email).toBeUndefined();
-  } else {
+  if (scenario === MalariaScenario.Trigger) {
     expect(response.body.activeEvents.email).toBeDefined();
+  } else {
+    expect(response.body.activeEvents.email).toBeUndefined();
   }
 
   expect(response.body.activeEvents.whatsapp).toBeFalsy();
@@ -53,13 +46,13 @@ export async function testFlashFloodScenario(
     (el) => (el as Element).textContent.toLowerCase(),
   ).map((el) => el.trim());
 
-  if (scenario === FlashFloodsScenario.NoTrigger) {
-    expect(eventNamesInEmail.length).toBe(0);
-  } else {
+  if (scenario === MalariaScenario.Trigger) {
     expect(eventNamesInEmail.length).toBe(eventNames.length);
+  } else {
+    expect(eventNamesInEmail.length).toBe(0);
   }
 
-  if (scenario !== FlashFloodsScenario.NoTrigger) {
+  if (scenario === MalariaScenario.Trigger) {
     // Check if each expected event name is included in at least one title
     for (const eventName of eventNames) {
       const eventTitle = getEventTitle(disasterTypeLabel, eventName);
