+++ tests/integration/tests/email/typhoon/email-phl-typhoon.test.ts
@@ -1,5 +1,5 @@
-import { TyphoonScenario } from '../../../../services/API-service/src/scripts/enum/mock-scenario.enum';
-import { getAccessToken, resetDB } from '../../helpers/utility.helper';
+import { TyphoonScenario } from '../../../../../services/API-service/src/scripts/enum/mock-scenario.enum';
+import { getAccessToken, resetDB } from '../../../helpers/utility.helper';
 import { testTyphoonScenario } from './test-typhoon-scenario.helper';
 
 const countryCodeISO3 = 'PHL';
