+++ tests/integration/tests/users/change-password.test.ts
@@ -1,10 +1,10 @@
-import { userData } from '../fixtures/users.const';
+import { userData } from '../../fixtures/users.const';
 import {
   changePassword,
   getAccessToken,
   loginUser,
   resetDB,
-} from '../helpers/utility.helper';
+} from '../../helpers/utility.helper';
 
 describe('change password of user ..', () => {
   let accessToken: string;
