+++ tests/integration/tests/users/create-user.test.ts
@@ -1,11 +1,11 @@
-import { UserRole } from '../../../services/API-service/src/api/user/user-role.enum';
-import { userData } from '../fixtures/users.const';
+import { UserRole } from '../../../../services/API-service/src/api/user/user-role.enum';
+import { userData } from '../../fixtures/users.const';
 import {
   createUser,
   getAccessToken,
   loginUser,
   resetDB,
-} from '../helpers/utility.helper';
+} from '../../helpers/utility.helper';
 
 describe('create user', () => {
   let accessToken: string;
@@ -15,7 +15,7 @@ describe('create user', () => {
     await resetDB(accessToken);
   });
 
-  it('successfully and log-in with it', async () => {
+  it('successfully, and log-in with it', async () => {
     // Arrange
     let newUserData = structuredClone(userData);
     newUserData.email = 'new-user@redcross.nl';
